requirements/prod.txt
package                   Version
------------------------- -------
asgiref                   3.3.1
Django                    3.1.4
mysql-connector           2.2.9
mysql-connector-python    8.0.22
mysql-connector-python-rf 2.2.2
mysqlclient               2.0.2
pip                       20.3.1
protobuf                  3.14.0
pytz                      2020.4
setuptools                51.0.0
six                       1.15.0
sqlparse                  0.4.1
wheel                     0.36.1
